+++ admin/db.inc.php
@@ -11,9 +11,9 @@ if ($tables_views && !$_POST["search"]) {
 		queries("SET foreign_key_checks = 0"); // allows to truncate or drop several tables at once
 	}
 
-	if ($_POST["truncate"] || $_POST["truncate_cascade"]) {
+	if ($_POST["truncate"]) {
 		if ($_POST["tables"]) {
-			$result = truncate_tables($_POST["tables"], (bool)$_POST["truncate_cascade"]);
+			$result = truncate_tables($_POST["tables"]);
 		}
 		$message = lang('Tables have been truncated.');
 	} elseif ($_POST["move"]) {
@@ -323,7 +323,6 @@ if ($_GET["ns"] === "") {
 			: "")))
 			. "<input type='submit' class='button' name='truncate' value='" . lang('Truncate') . "'> " .
 				help_script(DIALECT == "sqlite" ? "DELETE" : ("TRUNCATE" . (DIALECT == "pgsql" ? "" : " TABLE"))) . confirm()
-			. (DIALECT == "pgsql" ? "<input type='submit' class='button' name='truncate_cascade' value='" . lang('Truncate Cascade') . "'> " . help_script("TRUNCATE CASCADE") . confirm() : "")
 			. "<input type='submit' class='button' name='drop' value='" . lang('Drop') . "'>" . help_script("DROP TABLE") . confirm() . "\n";
 			$databases = (support("scheme") ? Admin::get()->getSchemas() : Admin::get()->getDatabases());
 			echo "</div></fieldset>\n";