+++ admin/drivers/sqlite.inc.php
@@ -861,7 +861,7 @@ if (isset($_GET["sqlite"])) {
 		return true;
 	}
 
-	function truncate_tables(array $tables, bool $cascade = false): bool
+	function truncate_tables(array $tables): bool
 	{
 		return apply_queries("DELETE FROM", $tables);
 	}