+++ admin/translations/_template.inc.php
@@ -171,7 +171,6 @@ return [
 	'Check' => 'Xx',
 	'Repair' => 'Xx',
 	'Truncate' => 'Xx',
-	'Truncate Cascade' => 'Xx',
 	'Tables have been truncated.' => 'Xx.',
 	'Move to other database' => 'Xx',
 	'Move' => 'Xx',