+++ CHANGELOG.md
@@ -21,6 +21,7 @@ AdminNeo 5.8.0
 - Database schema: Move table boxes also on touch screens
 - PostgreSQL: Hide table actions and row editing in pg_catalog and pg_toast
 - PostgreSQL: Order NULL last (by @vrana)
+- PostgreSQL: Remove the Truncate Cascade button (by @vrana)
 - MS SQL: Hide table actions and row editing in the sys schema
 - MongoDB, SimpleDB: Sort the list of tables by name
 - Drop compatibility with IIS 5 and with IIS 7.0 before its 2008 FastCGI update (by @vrana)